test: reorganize PagingDataHelperTest  to test core module
