+++ core/test/build.gradle.kts
@@ -7,4 +7,5 @@ android.namespace = "com.waffiq.bazz_movies.core.test"
 dependencies {
   implementation(libs.junit)
   implementation(libs.kotlinx.coroutines.test)
+  implementation(libs.recyclerview)
 }
