+++ feature/search/src/test/kotlin/com/waffiq/bazz_movies/feature/search/testutils/SearchTestVariables.kt
@@ -7,8 +7,8 @@ import com.waffiq.bazz_movies.core.common.utils.Constants.TV_MEDIA_TYPE
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.search.MultiSearchResponse
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.search.ResultsItemSearchResponse
 import com.waffiq.bazz_movies.feature.search.domain.model.ResultsItemSearch
-import com.waffiq.bazz_movies.feature.search.testutils.HelperTest.TestDiffCallback
-import com.waffiq.bazz_movies.feature.search.testutils.HelperTest.TestListCallback
+import com.waffiq.bazz_movies.core.test.PagingDataHelperTest.TestDiffCallback
+import com.waffiq.bazz_movies.core.test.PagingDataHelperTest.TestListCallback
 import kotlinx.coroutines.Dispatchers
 
 @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
